Create a production ready container
Run backend with uWSGI on production mode without development
requirements.

- Move dev-requirement* -> requirements-dev*
- Add uwsgi to requirements
- Add uWSGI configuration
- Run uWSGI from docker-entrypoint.sh
- Add staticbuilder and production stages to the Dockerfile
- Add ELASTICSEARCH_ANALYZER_MODE setting mainly for avoiding
  an error when collectstatic failed due to elasticsearch
  connection error.
- compose.yml use development image and make django depend
  on elasticsearch

Refs TIED-73
