+++ .docker/django/.env.example
@@ -16,7 +16,7 @@
 # --- docker-entrypoint.sh settings ---
 
 # Whether to run Django development server.
-DEV_SERVER=True
+# DEV_SERVER=True
 
 # Whether to create a default superuser when launching the container.
 # If set to True, the superuser will be created with the username and password
@@ -71,20 +71,20 @@ DATABASE_URL=postgres://root:root@helerm_postgres:5432/helerm
 # key for verifying that the token was meant for accessing this particular
 # HelERM instance (the tokens are signed, see below).
 # Does not correspond to standard Django setting
-#TOKEN_AUTH_ACCEPTED_AUDIENCE=string-identifying-this-helerm-instance
+# TOKEN_AUTH_ACCEPTED_AUDIENCE=string-identifying-this-helerm-instance
 
 # This key will be used by HelERM to verify the JWT token is from trusted
 # Identity Provider (OpenID terminology). The provider must have signed
 # the JWT TOKEN using this shared secret
 # Does not correspond to standard Django setting
-#TOKEN_AUTH_SHARED_SECRET=abcdefghacbdefgabcdefghacbdefgabcdefghacbdefgabcdefghacbdefgabcdefghacbdefg
+# TOKEN_AUTH_SHARED_SECRET=abcdefghacbdefgabcdefghacbdefgabcdefghacbdefgabcdefghacbdefgabcdefghacbdefg
 
 # Secret used for various functions within Django. This setting is
 # mandatory for Django, but HelERM will generate an ephemeral key,
 # if it is not defined here. Currently HelERM does not use any
 # functionality that needs this.
 # Django setting: SECRET_KEY https://docs.djangoproject.com/en/3.0/ref/settings/#secret-key
-#SECRET_KEY=gaehgoes89yrtp9384ygpe9r8ahgaerui8ghpae98rgh
+# SECRET_KEY=gaehgoes89yrtp9384ygpe9r8ahgaerui8ghpae98rgh
 
 # List of Host-values, that HelERM will accept in requests.
 # This setting is a Django protection measure against HTTP Host-header attacks
@@ -92,7 +92,7 @@ DATABASE_URL=postgres://root:root@helerm_postgres:5432/helerm
 # Specified as a comma separated list of allowed values. Note that this does
 # NOT matter if you are running with DEBUG
 # Django setting: ALLOWED_HOSTS https://docs.djangoproject.com/en/3.0/ref/settings/#allowed-hosts
-#ALLOWED_HOSTS=api.hel.ninja,helerm-api.hel.ninja
+ALLOWED_HOSTS=*
 
 # List of tuples (or just e-mail addresses) specifying Administrators of this
 # HelERM instance. Django uses this only when logging is configured to
@@ -121,7 +121,7 @@ COOKIE_PREFIX=helerm
 # like https://host.name/helerm). Note that this must include the
 # leading '/'.
 # Django setting: sets both CSRF_COOKIE_PATH and SESSION_COOKIE_PATH
-#PATH_PREFIX=/helerm
+# PATH_PREFIX=/helerm
 
 # Django INTERNAL_IPS setting allows some debugging aids for the addresses
 # specified here
@@ -133,30 +133,30 @@ INTERNAL_IPS=127.0.0.1
 # a proxy terminates the TLS connection and forwards the request over
 # a secure network. Specified using a tuple.
 # Django setting: SECURE_PROXY_SSL_HEADER https://docs.djangoproject.com/en/3.0/ref/settings/#secure-proxy-ssl-header
-#SECURE_PROXY_SSL_HEADER=('HTTP_X_FORWARDED_PROTO', 'https')
+# SECURE_PROXY_SSL_HEADER=('HTTP_X_FORWARDED_PROTO', 'https')
 
 # Media root is the place in file system where Django and, by extension
 # HelERM stores "uploaded" files. This means any and all files
 # that are inputted through API and generated by JHS191 exporter
 # Django setting: MEDIA_ROOT https://docs.djangoproject.com/en/3.0/ref/settings/#media-root
-#MEDIA_ROOT=/home/helerm/media
+# MEDIA_ROOT=/home/helerm/media
 
 # Static root is the place where HelERM will install any static
 # files that need to be served to clients. For HelERM this is mostly
 # JS and CSS for the API exploration interface + admin
 # Django setting: STATIC_ROOT
-#STATIC_ROOT=/home/helerm/static
+# STATIC_ROOT=/home/helerm/static
 
 # Media URL is address (URL) where users can access files in MEDIA_ROOT
 # through http. Ie. where your uploaded files are publicly accessible.
 # In the simple case this is a relative URL to same server as API
 # Django setting: MEDIA_URL https://docs.djangoproject.com/en/3.0/ref/settings/#media-url
-MEDIA_URL=/media/
+# MEDIA_URL=/media/
 
 # Static URL is address (URL) where users can access files in STATIC_ROOT
 # through http. Same factors apply as to MEDIA_URL
 # Django setting: STATIC_URL https://docs.djangoproject.com/en/3.0/ref/settings/#static-url
-STATIC_URL=/static/
+# STATIC_URL=/static/
 
 # Specifies that Django is to use `X-Forwarded-Host` as it would normally
 # use the `Host`-header. This is necessary when `Host`-header is used for
@@ -170,7 +170,7 @@ STATIC_URL=/static/
 # or purchased as PaaS. SENTRY_DSN setting specifies the URL where reports
 # for this HelERM instance should be sent. You can find this in
 # your Sentry interface (or through its API)
-#SENTRY_DSN=http://your.sentry.here/fsdafads/13
+# SENTRY_DSN=http://your.sentry.here/fsdafads/13
 
 # Sentry environment is an optional tag that can be included in sentry
 # reports. It is used to separate deployments within Sentry UI
@@ -178,6 +178,12 @@ SENTRY_ENVIRONMENT=local-development-unconfigured
 
 # Host for ElasticSearch connection.
 ELASTICSEARCH_HOST=helerm_elasticsearch:9200
+
+# Which analyzer will be used by elasticsearch. Default is to check if raudikko
+# plugin is available and use standard as a fallback. Probing can be skipped by setting
+# a specific analyzer.
+# Possible values are: probe, raudikko, standard
+# ELASTICSEARCH_ANALYZER_MODE = "probe"
 
 # Helsinki user management library (django-helusers) related OIDC settings.
 SOCIAL_AUTH_TUNNISTAMO_KEY=https://i/am/clientid/in/url/style
