+++ .docker/django/uwsgi_configuration.ini
@@ -1,0 +1,19 @@
+[uwsgi]
+# https://uwsgi-docs.readthedocs.io/en/latest/Options.html
+
+http-socket = :8000
+chdir = /app
+module = helerm.wsgi
+static-map = /static=/app/static
+uid = nobody
+gid = nogroup
+master = 1
+processes = 2
+threads = 2
+socket-timeout = 60
+
+# Suppress errors about clients closing sockets, happens when http pipes are closed
+# before workers have had the time to serve content to the pipe
+ignore-sigpipe = true
+ignore-write-errors = true
+disable-write-exception = true
