+++ .github/workflows/ci.yml
@@ -72,7 +72,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install --use-pep517 codecov -r requirements.txt -r dev-requirements.txt
+          pip install --use-pep517 codecov -r requirements.txt -r requirements-dev.txt
 
       - name: Configure sysctl limits
         run: |
