+++ requirements-dev.in
@@ -1,3 +1,5 @@
+-c requirements.txt
+
 black
 flake8
 flake8-bugbear
