+++ requirements-dev.txt
@@ -2,14 +2,14 @@
 # This file is autogenerated by pip-compile with Python 3.9
 # by the following command:
 #
-#    pip-compile --resolver=backtracking dev-requirements.in
+#    pip-compile --resolver=backtracking requirements-dev.in
 #
 attrs==22.2.0
     # via
     #   flake8-bugbear
     #   pytest
 black==23.1.0
-    # via -r dev-requirements.in
+    # via -r requirements-dev.in
 build==0.10.0
     # via pip-tools
 cfgv==3.3.1
@@ -28,19 +28,19 @@ filelock==3.9.0
     # via virtualenv
 flake8==6.0.0
     # via
-    #   -r dev-requirements.in
+    #   -r requirements-dev.in
     #   flake8-bugbear
     #   pep8-naming
 flake8-bugbear==23.2.13
-    # via -r dev-requirements.in
+    # via -r requirements-dev.in
 freezegun==1.2.2
-    # via -r dev-requirements.in
+    # via -r requirements-dev.in
 identify==2.5.19
     # via pre-commit
 iniconfig==2.0.0
     # via pytest
 isort==5.12.0
-    # via -r dev-requirements.in
+    # via -r requirements-dev.in
 mccabe==0.7.0
     # via flake8
 mypy-extensions==1.0.0
@@ -49,15 +49,16 @@ nodeenv==1.7.0
     # via pre-commit
 packaging==23.0
     # via
+    #   -c requirements.txt
     #   black
     #   build
     #   pytest
 pathspec==0.11.0
     # via black
 pep8-naming==0.13.3
-    # via -r dev-requirements.in
+    # via -r requirements-dev.in
 pip-tools==6.12.3
-    # via -r dev-requirements.in
+    # via -r requirements-dev.in
 platformdirs==3.1.0
     # via
     #   black
@@ -65,7 +66,7 @@ platformdirs==3.1.0
 pluggy==1.0.0
     # via pytest
 pre-commit==3.1.1
-    # via -r dev-requirements.in
+    # via -r requirements-dev.in
 pycodestyle==2.10.0
     # via flake8
 pyflakes==3.0.1
@@ -74,19 +75,23 @@ pyproject-hooks==1.0.0
     # via build
 pytest==7.2.2
     # via
-    #   -r dev-requirements.in
+    #   -r requirements-dev.in
     #   pytest-cov
     #   pytest-django
 pytest-cov==4.0.0
-    # via -r dev-requirements.in
+    # via -r requirements-dev.in
 pytest-django==4.5.2
-    # via -r dev-requirements.in
+    # via -r requirements-dev.in
 python-dateutil==2.8.2
-    # via freezegun
+    # via
+    #   -c requirements.txt
+    #   freezegun
 pyyaml==6.0
     # via pre-commit
 six==1.16.0
-    # via python-dateutil
+    # via
+    #   -c requirements.txt
+    #   python-dateutil
 tomli==2.0.1
     # via
     #   black
