+++ requirements.txt
@@ -150,3 +150,5 @@ urllib3==1.26.14
     #   elasticsearch
     #   requests
     #   sentry-sdk
+uwsgi==2.0.21
+    # via -r requirements.in
